Move unit tests out of leoserver.py into leo\unittests\core\test_leoserver.py
